** fixed warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.9@9412 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/SIMULATION/LABELING/ITRAQLabeler.C
@@ -349,7 +349,7 @@ namespace OpenMS
     DoubleReal offset = MS2_RT_time - elution_bounds[1];
     Int index = floor(offset / (width / (elution_ints.size()-1)) + 0.5);
 
-    OPENMS_POSTCONDITION(index < elution_ints.size(), "Wrong index computation! (Too large)")
+    OPENMS_POSTCONDITION(index < (Int)elution_ints.size(), "Wrong index computation! (Too large)")
 
     return elution_ints[index];
   }
